feat(service): Add metricas

+++ src/main/java/com/example/SalesHub/dto/request/EstoqueRequest.java
@@ -3,9 +3,10 @@ package com.example.SalesHub.dto.request;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
+import java.math.BigDecimal;
 
 @Builder
 public record EstoqueRequest(@Min(1) @NotNull
-                                                           Long produtoId,
-                                                           @Min(1) @NotNull
-                                                           Long quantidade) {}
+                             Long produtoId,
+                             @Min(1) @NotNull
+                             BigDecimal quantidade) {}
