+++ src/main/java/com/example/SalesHub/model/Venda.java
@@ -1,11 +1,11 @@
 package com.example.SalesHub.model;
 
-import com.example.SalesHub.model.enums.StatusPagamento;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.time.LocalDateTime;
@@ -21,7 +21,7 @@ public class Venda {
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "usuario",nullable = false)
+    @JoinColumn(name = "usuario", nullable = false)
     private Usuario usuario;
 
     @ManyToOne(fetch = FetchType.LAZY)
@@ -32,7 +32,7 @@ public class Venda {
     private BigDecimal valor;
 
     @Column(nullable = false)
-    private Long quantidade;
+    private BigDecimal quantidade;
 
     @Column(nullable = false)
     private BigDecimal valorTotalVendas;
